services/products/requirements.txt
fastapi
uvicorn
psycopg2
python-dotenv
alembic
fastapi_sqlalchemy